pyproject.toml
[project]
name = "openaq-dashboard"
version = "0.1.0"
description = "Dashboard for OpenAQ data leveraging Plotly Dash"
readme = "README.md"
requires-python = ">=3.12.9"
dependencies = [
    "boto3>=1.36.22",
    "dash>=2.18.2",
    "dash-ag-grid>=31.3.0",
    "dash-bootstrap-components>=1.7.1",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "isort>=6.0.0",
    "ruff>=0.9.6",
]
